#2 refactor structure to prepare new persist layer integration
